requirements.txt
torch~=1.11.0
torchvision~=0.12.0
gym[atari]~=0.21.0
pillow~=9.0.1
tensorboard
numpy~=1.21.3
future~=0.18.2
omegaconf~=2.1.1
ttkthemes
pandas~=1.4.2
seaborn~=0.11.2
tensorboard
hydra-core
